+++ node/full_node_integration_test.go
@@ -359,7 +359,7 @@ func (s *FullNodeTestSuite) TestGenesisInitialization() {
 
 	// Verify genesis state
 	state := s.node.blockManager.GetLastState()
-	require.Equal(uint64(s.node.genesis.InitialHeight()), state.InitialHeight)
+	require.Equal(s.node.genesis.InitialHeight(), state.InitialHeight)
 	require.Equal(s.node.genesis.ChainID(), state.ChainID)
 }
